service(refactor): add an AY connections_reader
